update bid engine
